refactor to prepare for Dockerfiles
